#10 react query ssr 적용
