+++ client/src/app/hydrateOnClient.tsx
@@ -1,0 +1,4 @@
+'use client';
+
+import { Hydrate as HydrateOnClient } from '@tanstack/react-query';
+export default HydrateOnClient;
