--- client/src/app/posts/getQueryClient.tsx
@@ -1,5 +1,0 @@
-import { QueryClient } from '@tanstack/react-query';
-import { cache } from 'react';
-
-const getQueryClient = cache(() => new QueryClient());
-export default getQueryClient;
