+++ client/src/components/Provider.tsx
@@ -1,26 +1,14 @@
 'use client';
 
-import React from 'react';
-import { QueryClientProvider, QueryClient } from '@tanstack/react-query';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { ReactNode } from 'react';
 
 type Props = {
-  children: React.ReactNode;
+  children: ReactNode;
 };
 
-function Providers({ children }: Props) {
-  const [client] = React.useState(
-    new QueryClient({
-      defaultOptions: {
-        // react-query 전역 설정
-        queries: {
-          refetchOnWindowFocus: false,
-          retry: false,
-        },
-      },
-    }),
-  );
+const queryClient = new QueryClient();
 
-  return <QueryClientProvider client={client}>{children}</QueryClientProvider>;
+export default function ReactQuery({ children }: Props) {
+  return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>;
 }
-
-export default Providers;
