+++ client/src/components/Recoil.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+import { ReactNode } from 'react';
+import { RecoilRoot } from 'recoil';
+
+type Props = {
+  children: ReactNode;
+};
+
+export default function Recoil({ children }: Props) {
+  return <RecoilRoot>{children}</RecoilRoot>;
+}
