+++ client/src/recoil/atom.tsx
@@ -1,10 +1,10 @@
-import { ThemeType } from '@/types/common';
-import { atom } from 'recoil';
-import { recoilPersist } from 'recoil-persist';
-const { persistAtom } = recoilPersist();
+// import { ThemeType } from '@/types/common';
+// import { atom } from 'recoil';
+// import { recoilPersist } from 'recoil-persist';
+// const { persistAtom } = recoilPersist();
 
-export const userThemeState = atom<ThemeType>({
-  key: 'userTheme',
-  default: 'dark',
-  effects_UNSTABLE: [persistAtom],
-});
+// export const userThemeState = atom<ThemeType>({
+//   key: 'userTheme',
+//   default: 'dark',
+//   effects_UNSTABLE: [persistAtom],
+// });
